+++ .gitignore
@@ -105,3 +105,6 @@ dist
 
 # pnpm
 .pnpm-store
+
+# storybook build output
+storybook-static/
